Added Final Cut Pro 10.3.4 Headers

+++ Headers/Frameworks/ProAppsFxSupport/PAEBlurOSC.h
@@ -8,7 +8,7 @@
 
 @interface PAEBlurOSC : PAEOSC
 {
-    long long _currentState;
+    unsigned int _currentState;
     unsigned int _offsetParameter;
 }
